+++ src/OFX.Reader.Application/OFX/Models/TransactionModel.cs
@@ -7,7 +7,7 @@ namespace OFX.Reader.Application.OFX.Models {
         public string TransactionType { get; set; }
         public DateTime TransactionDate { get; set; }
         public decimal TransactionAmount { get; set; }
-        public int TransactionId { get; set; }
+        public long TransactionId { get; set; }
         public string TransactionDescription { get; set; }
 
     }
